break out offset planning to its own file
