+++ go/vt/vtgate/planbuilder/operators/horizon_planning.go
@@ -163,125 +163,9 @@ func tryPushingDownProjection(
 	}
 }
 
-func planOffsets(ctx *plancontext.PlanningContext, root ops.Operator) (ops.Operator, error) {
-	visitor := func(in ops.Operator, _ semantics.TableSet, _ bool) (ops.Operator, rewrite.ApplyResult, error) {
-		var err error
-		switch op := in.(type) {
-		case *Horizon:
-			return nil, false, vterrors.VT13001("should not see Horizons here")
-		case *Derived:
-			return nil, false, vterrors.VT13001("should not see Derived here")
-		case *Filter:
-			err = planFilter(ctx, op)
-		case *Projection:
-			return planOffsetsForProjection(ctx, op)
-		case *ApplyJoin:
-			err = op.planOffsets(ctx)
-		}
-		if err != nil {
-			return nil, false, err
-		}
-		return in, rewrite.SameTree, nil
-	}
-
-	op, err := rewrite.BottomUp(root, TableID, visitor, stopAtRoute)
-	if err != nil {
-		if vterr, ok := err.(*vterrors.VitessError); ok && vterr.ID == "VT13001" {
-			// we encountered a bug. let's try to back out
-			return nil, errNotHorizonPlanned
-		}
-		return nil, err
-	}
-
-	return op, nil
-}
-
-func (a *ApplyJoin) planOffsets(ctx *plancontext.PlanningContext) error {
-	for _, col := range a.ColumnsAST {
-		// Read the type description for JoinColumn to understand the following code
-		for i, lhsExpr := range col.LHSExprs {
-			offset, err := a.pushColLeft(ctx, aeWrap(lhsExpr))
-			if err != nil {
-				return err
-			}
-			if col.RHSExpr == nil {
-				// if we don't have a RHS expr, it means that this is a pure LHS expression
-				a.Columns = append(a.Columns, -offset-1)
-			} else {
-				a.Vars[col.BvNames[i]] = offset
-			}
-		}
-		if col.RHSExpr != nil {
-			offset, err := a.pushColRight(ctx, aeWrap(col.RHSExpr))
-			if err != nil {
-				return err
-			}
-			a.Columns = append(a.Columns, offset+1)
-		}
-	}
-	return nil
-}
-
 func stopAtRoute(operator ops.Operator) rewrite.VisitRule {
 	_, isRoute := operator.(*Route)
 	return rewrite.VisitRule(!isRoute)
-}
-
-func planOffsetsForProjection(ctx *plancontext.PlanningContext, op *Projection) (ops.Operator, rewrite.ApplyResult, error) {
-	var err error
-	isSimple := true
-	for i, col := range op.Columns {
-		rewritten := sqlparser.CopyOnRewrite(col.GetExpr(), nil, func(cursor *sqlparser.CopyOnWriteCursor) {
-			col, ok := cursor.Node().(*sqlparser.ColName)
-			if !ok {
-				return
-			}
-			newSrc, offset, terr := op.Source.AddColumn(ctx, aeWrap(col))
-			if terr != nil {
-				err = terr
-			}
-			op.Source = newSrc
-			cursor.Replace(sqlparser.NewOffset(offset, col))
-		}, nil).(sqlparser.Expr)
-		if err != nil {
-			return nil, false, err
-		}
-
-		offset, ok := rewritten.(*sqlparser.Offset)
-		if ok {
-			// we got a pure offset back. No need to do anything else
-			op.Columns[i] = Offset{
-				Expr:   col.GetExpr(),
-				Offset: offset.V,
-			}
-			continue
-		}
-		isSimple = false
-
-		eexpr, err := evalengine.Translate(rewritten, nil)
-		if err != nil {
-			return nil, false, err
-		}
-
-		op.Columns[i] = Eval{
-			Expr:  rewritten,
-			EExpr: eexpr,
-		}
-	}
-	if !isSimple {
-		return op, rewrite.SameTree, nil
-	}
-
-	// is we were able to turn all the columns into offsets, we can use the SimpleProjection instead
-	sp := &SimpleProjection{
-		Source: op.Source,
-	}
-	for i, column := range op.Columns {
-		offset := column.(Offset)
-		sp.Columns = append(sp.Columns, offset.Offset)
-		sp.ASTColumns = append(sp.ASTColumns, &sqlparser.AliasedExpr{Expr: offset.Expr, As: sqlparser.NewIdentifierCI(op.ColumnNames[i])})
-	}
-	return sp, rewrite.NewTree, nil
 }
 
 func pushOrExpandHorizon(ctx *plancontext.PlanningContext, in *Horizon, isRoot bool) (ops.Operator, error) {
@@ -378,39 +262,6 @@ func expandHorizon(ctx *plancontext.PlanningContext, qp *QueryProjection, horizo
 			colName = expr.ColumnName()
 		}
 		proj.ColumnNames = append(proj.ColumnNames, colName)
-	}
-
-	return proj, nil
-}
-
-func pushProjections(ctx *plancontext.PlanningContext, qp *QueryProjection, src ops.Operator, isRoot bool) (ops.Operator, error) {
-	// if we are at the root, we have to return the columns the user asked for. in all other levels, we reuse as much as possible
-	var proj *SimpleProjection
-	if isRoot {
-		proj = newSimpleProjection(src)
-	}
-	for _, e := range qp.SelectExprs {
-		expr, err := e.GetAliasedExpr()
-		if err != nil {
-			return nil, err
-		}
-		if !expr.As.IsEmpty() {
-			// we are not handling column names correct yet, so let's fail here for now
-			return nil, errNotHorizonPlanned
-		}
-		var offset int
-		src, offset, err = src.AddColumn(ctx, expr)
-		if err != nil {
-			return nil, err
-		}
-		if isRoot {
-			proj.ASTColumns = append(proj.ASTColumns, expr)
-			proj.Columns = append(proj.Columns, offset)
-		}
-	}
-
-	if !isRoot {
-		return src, nil
 	}
 
 	return proj, nil
